+++ webapp/src/components/OnSaleList/OnSaleListItem/OnSaleListItem.types.ts
@@ -1,0 +1,3 @@
+import { Item } from '../OnSaleList.types'
+
+export type Props = { item: Item }
