+++ webapp/src/components/OnSaleList/OnSaleListItem/index.ts
@@ -1,0 +1,3 @@
+import OnSaleListItem from './OnSaleListItem'
+
+export default OnSaleListItem
